requirements.txt
pyrogram
tgcrypto
pyaes==1.6.1
pysocks==1.7.1
aiosqlite>=0.16.0,<0.18.0
python-dotenv
